refactor : 피드백 반영

+++ src/main/java/com/example/puddingbe/domain/ranking/domain/Ranking.java
@@ -18,5 +18,5 @@ public class Ranking {
     private Long userId;
 
     @Column(name = "total_pudding", nullable = false)
-    private int totalPudding;
+    private long totalPudding;
 }
