+++ src/main/java/com/example/puddingbe/domain/ranking/presentation/dto/RankingResponse.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 public class RankingResponse {
     private long id;
     private long userId;
-    private int totalPudding;
+    private long totalPudding;
 }
